+++ Tests/Integration/MSFT_xIPAddress.Integration.Tests.ps1
@@ -31,23 +31,32 @@ try
 
     Describe "$($script:DSCResourceName)_Integration" {
         #region DEFAULT TESTS
-        It 'Should compile without throwing' {
+        It 'Should compile and apply the MOF without throwing' {
             {
                 & "$($script:DSCResourceName)_Config" -OutputPath $TestDrive
-                Start-DscConfiguration -Path $TestDrive -ComputerName localhost -Wait -Verbose -Force
-            } | Should not throw
+
+                Start-DscConfiguration `
+                    -Path $TestDrive `
+                    -ComputerName localhost `
+                    -Wait `
+                    -Verbose `
+                    -Force `
+                    -ErrorAction Stop
+            } | Should Not Throw
         }
 
         It 'should be able to call Get-DscConfiguration without throwing' {
-            { Get-DscConfiguration -Verbose -ErrorAction Stop } | Should Not throw
+            { Get-DscConfiguration -Verbose -ErrorAction Stop } | Should Not Throw
         }
         #endregion
 
         It 'Should have set the resource and all the parameters should match' {
-            $current = Get-DscConfiguration | Where-Object {$_.ConfigurationName -eq "$($script:DSCResourceName)_Config"}
-            $current.InterfaceAlias             | Should Be $TestIPAddress.InterfaceAlias
-            $current.AddressFamily              | Should Be $TestIPAddress.AddressFamily
-            $current.IPAddress                  | Should Be $TestIPAddress.IPAddress
+            $current = Get-DscConfiguration | Where-Object -FilterScript {
+                $_.ConfigurationName -eq "$($script:DSCResourceName)_Config"
+            }
+            $current.InterfaceAlias | Should Be $TestIPAddress.InterfaceAlias
+            $current.AddressFamily  | Should Be $TestIPAddress.AddressFamily
+            $current.IPAddress      | Should Be $TestIPAddress.IPAddress
         }
     }
     #endregion